Clean up after removal of Scala 2.12 support
