--- src/main/scala-2.12/com.scalatsi/DefaultTSTypes.scala
@@ -1,13 +1,0 @@
-package com.scalatsi
-
-/** The default Scala-to-Typescript mappings provided by this library */
-// Cannot be put in the scala-agnostic code because 2.12 does not support literal types
-trait DefaultTSTypes
-    extends PrimitiveTSTypes
-    with ScalaTSTypes
-    with CollectionTSTypes
-    with TupleTSTypes
-    with FunctionTSTypes
-    with JavaTSTypes
-    with ScalaEnumTSTypes {}
-object DefaultTSTypes extends DefaultTSTypes
