+++ src/main/scala/com/scalatsi/LanguageTSTypes.scala
@@ -4,7 +4,6 @@ import TypescriptType._
 
 import scala.collection.immutable.ListMap
 import scala.reflect.ClassTag
-import scala.reflect.runtime.universe.TypeTag
 
 trait ScalaTSTypes {
   implicit val anyTSType: TSType[Any]   = TSType(TSAny)
@@ -34,25 +33,10 @@ trait LowPriorityCollectionTSType {
 }
 
 trait ScalaEnumTSTypes {
-  implicit def scalaEnumTSType[E <: Enumeration: TypeTag]: TSNamedType[E] = {
-    // we have to use reflection to call the enum values() method, no easy way to get the object directly in Scala 2.12
-    // might be possible to improve using 2.13+ singleton types
-    val tpe  = scala.reflect.runtime.universe.typeOf[E]
-    val name = tpe.typeSymbol.name.toString
-    val members = tpe.members.filter(sym =>
-      // For some reason the types here are not subtypes of E#Value or Enumeration#Value (=:= and <:< return false)
-      // Instead filter out everything but "private[this] val"s. Those shouldn't be in Enumeration's anyway
-      sym.isPrivateThis &&
-        !sym.isMethod &&
-        !sym.isClass &&
-        !sym.isType &&
-        !sym.isModule
-    )
-
-    val values = members
-      .map(_.name.toString.trim)
-      .toSeq
-      .sorted
+  implicit def scalaEnumTSType[E <: Enumeration: ValueOf]: TSNamedType[E] = {
+    val e                   = valueOf[E]
+    val name: String        = e.toString()
+    val values: Seq[String] = e.values.map(id => id.toString).toSeq
 
     TSType.alias[E](name, TSUnion(values.map(TSLiteralString.apply)))
   }
@@ -158,4 +142,22 @@ trait FunctionTSTypes {
     TSType(TSFunction(ListMap("arg0" -> t1.get, "arg1" -> t2.get, "arg2" -> t3.get, "arg3" -> t4.get), r.get))
 
   // TODO: Function5-22
+}
+
+trait LiteralTSTypes {
+
+  implicit val tsLiteralTrue: TSType[true]   = TSType(TSLiteralBoolean(true))
+  implicit val tsLiteralFalse: TSType[false] = TSType(TSLiteralBoolean(false))
+
+  implicit def tsLiteralString[T <: Singleton with String: ValueOf]: TSType[T] = TSType(TSLiteralString(valueOf[T]))
+
+  private def number[T](bd: BigDecimal): TSType[T] = TSType(TSLiteralNumber(bd))
+
+  implicit def tsLiteralInt[T <: Singleton with Int: ValueOf]: TSType[T] = number(valueOf[T]: Int)
+
+  implicit def tsLiteralLong[T <: Singleton with Long: ValueOf]: TSType[T] = number(valueOf[T]: Long)
+
+  implicit def tsLiteralDouble[T <: Singleton with Double: ValueOf]: TSType[T] = number(valueOf[T]: Double)
+
+  implicit def tsLiteralFloat[T <: Singleton with Float: ValueOf]: TSType[T] = number(valueOf[T]: Float)
 }
